Authentication System

+++ routes/web.php
@@ -21,15 +21,13 @@ Route::get('/home', function () {
     return view('home');
 });
 
-Route::get('/login', function () {
-    return view('login');
-});
+Route::get('/register', 'AuthController@getRegister')->name('register')->middleware('guest');
+Route::post('/register', 'AuthController@postRegister')->middleware('guest');
+Route::get('/login', 'AuthController@getLogin')->name('login')->middleware('guest');
+Route::post('/login', 'AuthController@postLogin')->middleware('guest');
 
 Route::get('/info', 'InfoController@info');
 
-Route::get('/register', function () {
-    return view('register');
-});
 
 Route::get('/about', function () {
     return view('about');
@@ -44,20 +42,16 @@ Route::get('/detailBerita', function () {
 
 Route::get('/dashboardAdmin', function () {
     return view('dashboardAdmin');
-});
+})->middleware('auth')->name('dashboardAdmin');
+Route::get('/logout', 'AuthController@logout')->middleware('auth')->name('logout');
 
-Route::get('/beritaAdmin', 'BeritaController@index');
-Route::get('/beritaAdmin/create', 'BeritaController@create');
-Route::post('/beritaAdmin', 'BeritaController@store');
-Route::get('/beritaAdmin/{id}/edit', 'BeritaController@edit');
-Route::patch('/beritaAdmin/{id}', 'BeritaController@update');
-Route::delete('/beritaAdmin/{id}', 'BeritaController@destroy');
+Route::get('/beritaAdmin', 'BeritaController@index')->middleware('auth')->name('beritaAdmin');
+Route::get('/beritaAdmin/create', 'BeritaController@create')->middleware('auth')->name('beritaAdminCreate');
+Route::post('/beritaAdmin', 'BeritaController@store')->middleware('auth')->name('beritaAdminstore');
+Route::get('/beritaAdmin/{id}/edit', 'BeritaController@edit')->middleware('auth')->name('beritaAdminedit');
+Route::patch('/beritaAdmin/{id}', 'BeritaController@update')->middleware('auth')->name('beritaAdminupdate');
+Route::delete('/beritaAdmin/{id}', 'BeritaController@destroy')->middleware('auth')->name('beritaAdmindestroy');
 
-Route::get('/hospitalAdmin', function () {
-    return view('hospitalAdmin');
-});
-Route::get('/formHospital', function () {
-    return view('formHospital');
-});
-Route::get('/hospitalAdmin', 'RujukController@index');
-Route::get('/formHospital', 'RujukController@create');
+
+Route::get('/hospitalAdmin', 'RujukController@index')->middleware('auth')->name('hospitalAdmin');
+Route::get('/formHospital', 'RujukController@create')->middleware('auth')->name('formAdmin');
